+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.19.0] - 2023-05-10
+
 ### Added
 
 - Added function to infer filetype when loading spectra
@@ -478,7 +480,8 @@ Contains three plot types: `plot_spectrum()` or `spectrum.plot()`, `plot_spectra
 - This is the initial version of Spec2Vec from https://github.com/iomega/Spec2Vec
 - (later splitted into matchms + spec2vec)
 
-[Unreleased]: https://github.com/matchms/matchms/compare/0.18.0...HEAD
+[Unreleased]: https://github.com/matchms/matchms/compare/0.19.0...HEAD
+[0.19.0]: https://github.com/matchms/matchms/compare/0.18.0...0.19.0
 [0.18.0]: https://github.com/matchms/matchms/compare/0.17.0...0.18.0
 [0.17.0]: https://github.com/matchms/matchms/compare/0.16.0...0.17.0
 [0.16.0]: https://github.com/matchms/matchms/compare/0.15.0...0.16.0